[FLINK-2196] [javaAPI] Moved misplaced SortPartitionOperator class
This closes #817
